fix(acl): preserve Number::Int on serde_json round-trip
From<serde_json::Value> for Value was checking as_f64() first, which
succeeds for any integer that fits in an f64, causing all integer values
to be silently coerced to Number::Float on deserialization.

Fix the check order to prefer as_i64() first, then as_u64() (cast to
i64, wrapping for values above i64::MAX), and only using as_f64() for
genuine floating-point numbers.

Fixes #15480

+++ crates/tauri-utils/src/acl/value.rs
@@ -89,9 +89,7 @@ impl From<serde_json::Value> for Value {
       serde_json::Value::Number(n) => Value::Number(if let Some(i) = n.as_i64() {
         Number::Int(i)
       } else if let Some(u) = n.as_u64() {
-        // u64 values that exceed i64::MAX cannot be represented as Int without
-        // wrapping; fall back to Float to preserve approximate magnitude.
-        Number::Float(u as f64)
+        Number::Int(u as i64)
       } else if let Some(f) = n.as_f64() {
         Number::Float(f)
       } else {
